Allow switching languages midword
Allows switching languages midword.
When in predictive mode, a new match is searched for based on the previously
entered keys, allowing for quick corrections between languages.
When in ABC mode, the current letter is not committed, and the letter list is
updated.

Since the status bar is no longer visible when typing, immediate feedback is
shown by using a toast.